@@ -91,10 +91,10 @@
     "lodash-webpack-plugin": "^0.11.5",
     "prettier": "^1.19.1",
     "prettier-eslint-cli": "^5.0.0",
-    "react": "~16.4.2",
-    "react-dom": "~16.4.2",
+    "react": "~17.0.2",
+    "react-dom": "~17.0.2",
     "react-redux": "~6.0.1",
-    "react-test-renderer": "^16.13.1",
+    "react-test-renderer": "^17.0.2",
     "redux": "~3.7.2",
     "redux-immutable": "^4.0.0",
     "rimraf": "^2.7.1",
